requirements.txt
Scrapy==2.11.2
selenium==4.25.0
requests==2.32.3
PyYAML==6.0.2
